UPDATE [sc-1049] - refactored templater service
