+++ apps/course-authoring/src/main/services/index.ts
@@ -3,3 +3,4 @@ export * as FileSystem from './file-system';
 export * as InternalStorage from './internal-storage';
 export * as Menu from './menu';
 export * as Requester from './requester';
+export * as Templater from './templater';
